fix: modify folder structure for types object
