+++ src/mcpm/core/mcp/types.py
@@ -1,0 +1,8 @@
+from enum import Enum
+
+
+class ResourceType(str, Enum):
+    TOOL = "tool"
+    PROMPT = "prompt"
+    RESOURCE = "resource"
+    RESOURCE_TEMPLATE = "resource_template"
