+++ src/mcpm/core/router/manager.py
@@ -17,7 +17,8 @@ from mcp.types import (
 from pydantic import AnyUrl
 
 from mcpm.core.mcp.client_connection import ServerConnection
-from mcpm.core.schema import ResourceType, ServerConfig
+from mcpm.core.mcp.types import ResourceType
+from mcpm.core.schema import ServerConfig
 from mcpm.utils.config import PROMPT_SPLITOR, RESOURCE_SPLITOR, TOOL_SPLITOR
 from mcpm.utils.errlog_manager import ServerErrorLogManager
